# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SimpleRecipes",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/recipes": {
      "get": {
        "summary": "Get recipes",
        "responses": {
          "200": {
            "description": "List of recipes on the server",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "description": "The recipes",
                  "items": {
                    "$ref": "#/components/schemas/Recipe"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create recipe",
        "requestBody": {
          "description": "The recipe to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recipe successfully created"
          }
        }
      }
    },
    "/recipes/{id}": {
      "get": {
        "summary": "Get recipe by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "description": "The ID of the recipe to get"
          }
        ],
        "responses": {
          "200": {
            "description": "The retrieved recipe",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The retrieved recipe",
                  "$ref": "#/components/schemas/Recipe"
                }
              }
            }
          }
        }
      }
    },
    "/recipes/{id}/steps": {
      "post": {
        "summary": "Create a recipe step",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "description": "The ID of the recipe to create a new step for"
          }
        ],
        "requestBody": {
          "description": "The step to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecipeStepBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recipe step created successfully"
          }
        }
      }
    },
    "/ingredients": {
      "get": {
        "summary": "Get ingredients",
        "responses": {
          "200": {
            "description": "The list of ingredients on the server",
            "content": {
              "application/json": {
                "schema": {
                  "description": "List of ingredients",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ingredient"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create ingredient",
        "requestBody": {
          "description": "The ingredient to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIngredientBody"
              }
            }
          }
        }
      }
    },
    "/ingredients/{id}": {
      "put": {
        "summary": "Update ingredient",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "description": "The ID of the ingredient to update"
          }
        ],
        "requestBody": {
          "description": "The new ingredient data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIngredientBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successfully updated ingredient"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Recipe": {
        "description": "A cooking recipe",
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseRecipe"
          },
          {
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "description": "The unique identifier of the recipe"
              },
              "steps": {
                "type": "array",
                "description": "The list of steps in this recipe",
                "items": {
                  "$ref": "#/components/schemas/RecipeStep"
                }
              },
              "ingredients": {
                "type": "array",
                "description": "The list of ingredients used in this recipe",
                "items": {
                  "$ref": "#/components/schemas/RecipeIngredient"
                }
              }
            },
            "required": [
              "id",
              "steps",
              "ingredients"
            ]
          }
        ]
      },
      "RecipeIngredient": {
        "description": "An ingredient used in a recipe",
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseRecipeIngredient"
          },
          {
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "description": "The unique identifier of the recipe ingredient"
              },
              "ingredient": {
                "$ref": "#/components/schemas/Ingredient"
              }
            },
            "required": [
              "id",
              "ingredient"
            ]
          }
        ]
      },
      "IngredientUnit": {
        "type": "string",
        "description": "The unit of the ingredient amount"
      },
      "Ingredient": {
        "type": "object",
        "description": "An ingredient used in recipes",
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseIngredient"
          },
          {
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "description": "The unique identifier for the ingredient"
              }
            },
            "required": [
              "id"
            ]
          }
        ]
      },
      "RecipeStep": {
        "description": "A step in a cooking recipe",
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseRecipeStep"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "description": "The unique identifier of the recipe step"
              },
              "ingredients": {
                "type": "array",
                "description": "The list of identifiers of the ingredients used by this step",
                "items": {
                  "type": "string",
                  "format": "uuid",
                  "description": "An identifier referencing an item in the ingredients array of Recipe objects"
                }
              }
            },
            "required": [
              "id",
              "ingredients"
            ]
          }
        ]
      },
      "CreateRecipeBody": {
        "type": "object",
        "description": "A recipe sent by the client to create",
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseRecipe"
          },
          {
            "properties": {
              "steps": {
                "type": "array",
                "description": "The list of steps in this recipe",
                "items": {
                  "description": "A step in a cooking recipe",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseRecipeStep"
                    },
                    {
                      "properties": {
                        "ingredients": {
                          "type": "array",
                          "description": "The list of identifiers of the ingredients used by this step",
                          "items": {
                            "type": "string",
                            "description": "An identifier referencing an item in the ingredients array of Recipe objects"
                          }
                        }
                      },
                      "required": [
                        "ingredients"
                      ]
                    }
                  ]
                }
              },
              "ingredients": {
                "type": "array",
                "description": "The list of ingredients used in this recipe",
                "items": {
                  "description": "An ingredient used in this recipe",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BaseRecipeIngredient"
                    },
                    {
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "The identifier of the recipe ingredient (unique in this request)"
                        },
                        "ingredientId": {
                          "type": "string",
                          "format": "uuid",
                          "description": "An identifier referencing an Ingredient"
                        }
                      },
                      "required": [
                        "id",
                        "ingredientId"
                      ]
                    }
                  ]
                }
              }
            },
            "required": [
              "title",
              "description",
              "steps",
              "ingredients"
            ]
          }
        ]
      },
      "CreateRecipeStepBody": {
        "type": "object",
        "description": "A recipe step sent by the client to create",
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseRecipeStep"
          },
          {
            "properties": {
              "ingredients": {
                "type": "array",
                "description": "The list of identifiers of the ingredients used by this step",
                "items": {
                  "type": "string",
                  "format": "uuid",
                  "description": "An identifier referencing an item in the ingredients array of Recipe objects"
                }
              }
            },
            "required": [
              "ingredients"
            ]
          }
        ]
      },
      "CreateIngredientBody": {
        "type": "object",
        "description": "An ingredient sent by the client to create it",
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseIngredient"
          }
        ]
      },
      "BaseRecipe": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The name of the recipe"
          },
          "description": {
            "type": "string",
            "description": "The description of the recipe"
          }
        },
        "required": [
          "title",
          "description"
        ]
      },
      "BaseRecipeIngredient": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "description": "The amount of the ingredient used in the recipe"
          },
          "unit": {
            "$ref": "#/components/schemas/IngredientUnit"
          }
        },
        "required": [
          "quantity",
          "unit"
        ]
      },
      "BaseRecipeStep": {
        "type": "object",
        "properties": {
          "instruction": {
            "type": "string",
            "description": "The user-readable instructions for this step"
          },
          "ordinal": {
            "type": "number",
            "description": "The order of this step within the recipe"
          }
        },
        "required": [
          "instruction",
          "ordinal"
        ]
      },
      "BaseIngredient": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The user-readable name of the ingredient"
          },
          "defaultUnit": {
            "type": "string",
            "description": "The default unit for the ingredient"
          }
        },
        "required": [
          "name",
          "defaultUnit"
        ]
      }
    }
  }
}